requirements.txt
imap-tools
requests
beautifulsoup4
python-dotenv